refactor(dashboard): modularize spec-create-modal into custom hook and presentational subcomponents
