+++ tools/dashboard/src/components/spec-create-modal/index.tsx
@@ -1,0 +1,5 @@
+export { SpecCreateModal, type SpecCreateModalProps } from './spec-create-modal';
+export { useSpecCreateForm } from './use-spec-create-form';
+export { SpecMetadataFields } from './spec-metadata-fields';
+export { SpecAiPlanningSection } from './spec-ai-planning-section';
+export { SpecCreateErrorBanner } from './spec-create-error-banner';
